add more functions with test cases

+++ srcs/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: jpollore <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/19 18:45:54 by jpollore          #+#    #+#             */
-/*   Updated: 2018/02/22 19:22:10 by jpollore         ###   ########.fr       */
+/*   Updated: 2018/02/22 20:07:39 by jpollore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (str[count])
